Refactor: 설정 파일 위치 변경 [#2]
