+++ src/main/java/gg/loto/common/config/JpaConfig.java
@@ -1,4 +1,4 @@
-package gg.loto.common;
+package gg.loto.common.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
